feat: Add condition field to Painting model and set default value
Add a new field called "condition" to the Painting model in order to track the condition of each painting. The field is required and has a default value of 100. This change ensures that the condition of each painting is properly recorded and can be used for further analysis or filtering.